Socket Constructor; iostream; debug statements
A constructor that takes a socket has been added, but not implemented.
This will be likey be used after an accept call has returned to create a
new instance of this class for another thread.

iostream has been added to use cout and cin, which is being used in ome
of the debug statements.